+++ frameworks/CSharp/appmpower/odbcinst.ini
@@ -16,9 +16,9 @@ Description=ODBC for PostgreSQL
 ; path than /usr/local/lib/ following your installation.
 ; This is the standard location used by apt-get install -y unixodbc
 ;ON SERVER
-Driver = /usr/lib/x86_64-linux-gnu/odbc/psqlodbcw.so
+;Driver = /usr/lib/x86_64-linux-gnu/odbc/psqlodbcw.so
 ;TODOLOCAL: ON MAC
-;Driver =/usr/lib/aarch64-linux-gnu/odbc/psqlodbcw.so
+Driver =/usr/lib/aarch64-linux-gnu/odbc/psqlodbcw.so
 
 ;Driver =/usr/local/pgsqlodbc/lib/psqlodbcw.so
 Threading = 0
