+++ frameworks/CSharp/appmpower/src/appMpower.Orm/Data/DbCommand.cs
@@ -1,7 +1,7 @@
 using System.Data;
 using System.Threading.Tasks;
 
-namespace appMpowerAot.Data
+namespace appMpower.Orm.Data
 {
    public class DbCommand : IDbCommand
    {
