+++ frameworks/CSharp/appmpower/src/appMpower.Orm/DotnetMethods.cs
@@ -1,10 +1,10 @@
 using System.Runtime.InteropServices;
 using System.Text; 
 using System.Text.Json; 
-using appMpowerAot.DataObjects; 
-using appMpowerAot.Serializers; 
+using appMpower.Orm.Objects; 
+using appMpower.Orm.Serializers; 
 
-namespace appMpowerAot;
+namespace appMpower.Orm;
 
 public static class DotnetMethods
 {
