+++ frameworks/CSharp/appmpower/src/appMpower.Orm/Microsoft/BatchUpdateString.cs
@@ -2,8 +2,8 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information. 
 
 using System.Linq;
-using appMpowerAot; 
-using appMpowerAot.Data; 
+using appMpower.Orm; 
+using appMpower.Orm.Data; 
 
 namespace PlatformBenchmarks
 {
