+++ frameworks/CSharp/appmpower/src/appMpower.Orm/Microsoft/CachKey.cs
