+++ frameworks/CSharp/appmpower/src/appMpower.Orm/Microsoft/StringBuilderCache.cs
