+++ frameworks/CSharp/appmpower/src/appMpower.Orm/Objects/Fortune.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace appMpowerAot.DataObjects
+namespace appMpower.Orm.Objects
 {
    public readonly struct Fortune : IComparable<Fortune>, IComparable
    {
