+++ frameworks/CSharp/appmpower/src/appMpower.Orm/Objects/World.cs
@@ -1,4 +1,4 @@
-namespace appMpowerAot.DataObjects
+namespace appMpower.Orm.Objects
 {
    public struct World
    {
