+++ frameworks/CSharp/appmpower/src/appMpower.Orm/appMpower.Orm.csproj
@@ -8,9 +8,9 @@
     <SelfContained>true</SelfContained>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     
-    <RuntimeIdentifier>linux-x64</RuntimeIdentifier> <!-- docker server -->
+    <!--<RuntimeIdentifier>linux-x64</RuntimeIdentifier>--> <!-- docker server -->
     <!--TODOLOCAL-->
-    <!--<RuntimeIdentifier>linux-arm64</RuntimeIdentifier>--> <!-- docker local -->
+    <RuntimeIdentifier>linux-arm64</RuntimeIdentifier> <!-- docker local -->
     <!--<RuntimeIdentifier>osx-arm64</RuntimeIdentifier>--> <!-- outside docker local -->
 
     <InvariantGlobalization>true</InvariantGlobalization>
