+++ frameworks/CSharp/appmpower/src/appMpower/appMpower.csproj
@@ -10,15 +10,13 @@
     <None Include="appsettings.json" CopyToOutputDirectory="PreserveNewest" />
 
     <!--TODOLOCAL-->
-    <!--
-    <Content Include="../appMpowerAot/bin/Release/net8.0/osx-arm64/native/appMpowerAot.dylib">
+    <Content Include="../appMpower.Orm/bin/Release/net8.0/osx-arm64/native/appMpower.Orm.dylib">
         <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content> 
-    -->
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include = "..\appMpowerAot\appMpowerAot.csproj" />
+    <ProjectReference Include = "..\appMpower.Orm\appMpower.Orm.csproj" />
   </ItemGroup>
 
   <PropertyGroup>
